refactor: move dll related code to dll directory
