+++ lib/dll/DelegatedPlugin.js
@@ -5,10 +5,10 @@
 
 "use strict";
 
+const DelegatedSourceDependency = require("../dependencies/DelegatedSourceDependency");
 const DelegatedModuleFactoryPlugin = require("./DelegatedModuleFactoryPlugin");
-const DelegatedSourceDependency = require("./dependencies/DelegatedSourceDependency");
 
-/** @typedef {import("./Compiler")} Compiler */
+/** @typedef {import("../Compiler")} Compiler */
 /** @typedef {import("./DelegatedModuleFactoryPlugin").Options} Options */
 
 const PLUGIN_NAME = "DelegatedPlugin";
