+++ lib/dll/DllEntryPlugin.js
@@ -5,12 +5,12 @@
 
 "use strict";
 
+const DllEntryDependency = require("../dependencies/DllEntryDependency");
+const EntryDependency = require("../dependencies/EntryDependency");
 const DllModuleFactory = require("./DllModuleFactory");
-const DllEntryDependency = require("./dependencies/DllEntryDependency");
-const EntryDependency = require("./dependencies/EntryDependency");
 
-/** @typedef {import("./Compiler")} Compiler */
-/** @typedef {import("./Entrypoint").EntryOptions} EntryOptions */
+/** @typedef {import("../Compiler")} Compiler */
+/** @typedef {import("../Entrypoint").EntryOptions} EntryOptions */
 
 /** @typedef {string[]} Entries */
 /** @typedef {EntryOptions & { name: string }} Options */
