+++ lib/dll/DllModuleFactory.js
@@ -5,12 +5,12 @@
 
 "use strict";
 
+const ModuleFactory = require("../ModuleFactory");
 const DllModule = require("./DllModule");
-const ModuleFactory = require("./ModuleFactory");
 
-/** @typedef {import("./ModuleFactory").ModuleFactoryCallback} ModuleFactoryCallback */
-/** @typedef {import("./ModuleFactory").ModuleFactoryCreateData} ModuleFactoryCreateData */
-/** @typedef {import("./dependencies/DllEntryDependency")} DllEntryDependency */
+/** @typedef {import("../ModuleFactory").ModuleFactoryCallback} ModuleFactoryCallback */
+/** @typedef {import("../ModuleFactory").ModuleFactoryCreateData} ModuleFactoryCreateData */
+/** @typedef {import("../dependencies/DllEntryDependency")} DllEntryDependency */
 
 class DllModuleFactory extends ModuleFactory {
 	constructor() {
