+++ lib/dll/DllPlugin.js
@@ -5,12 +5,12 @@
 
 "use strict";
 
+const FlagAllModulesAsUsedPlugin = require("../FlagAllModulesAsUsedPlugin");
 const DllEntryPlugin = require("./DllEntryPlugin");
-const FlagAllModulesAsUsedPlugin = require("./FlagAllModulesAsUsedPlugin");
 const LibManifestPlugin = require("./LibManifestPlugin");
 
-/** @typedef {import("../declarations/plugins/DllPlugin").DllPluginOptions} DllPluginOptions */
-/** @typedef {import("./Compiler")} Compiler */
+/** @typedef {import("../../declarations/plugins/DllPlugin").DllPluginOptions} DllPluginOptions */
+/** @typedef {import("../Compiler")} Compiler */
 /** @typedef {import("./DllEntryPlugin").Entries} Entries */
 /** @typedef {import("./DllEntryPlugin").Options} Options */
 
@@ -34,13 +34,13 @@ class DllPlugin {
 	apply(compiler) {
 		compiler.hooks.validate.tap(PLUGIN_NAME, () => {
 			compiler.validate(
-				() => require("../schemas/plugins/DllPlugin.json"),
+				() => require("../../schemas/plugins/DllPlugin.json"),
 				this.options,
 				{
 					name: "Dll Plugin",
 					baseDataPath: "options"
 				},
-				(options) => require("../schemas/plugins/DllPlugin.check")(options)
+				(options) => require("../../schemas/plugins/DllPlugin.check")(options)
 			);
 		});
