+++ lib/dll/DllReferencePlugin.js
@@ -5,20 +5,19 @@
 
 "use strict";
 
-const DelegatedModuleFactoryPlugin = require("./DelegatedModuleFactoryPlugin");
 const ExternalModuleFactoryPlugin = require("./ExternalModuleFactoryPlugin");
-const DelegatedSourceDependency = require("./dependencies/DelegatedSourceDependency");
 const WebpackError = require("./errors/WebpackError");
+const DelegatedSourceDependency = require("./dependencies/DelegatedSourceDependency");
 const { makePathsRelative } = require("./util/identifier");
 const parseJson = require("./util/parseJson");
 
-/** @typedef {import("../declarations/WebpackOptions").Externals} Externals */
-/** @typedef {import("../declarations/plugins/DllReferencePlugin").DllReferencePluginOptions} DllReferencePluginOptions */
-/** @typedef {import("../declarations/plugins/DllReferencePlugin").DllReferencePluginOptionsContent} DllReferencePluginOptionsContent */
-/** @typedef {import("../declarations/plugins/DllReferencePlugin").DllReferencePluginOptionsManifest} DllReferencePluginOptionsManifest */
-/** @typedef {import("./Compiler")} Compiler */
-/** @typedef {import("./Compiler").CompilationParams} CompilationParams */
-/** @typedef {import("./util/fs").InputFileSystem} InputFileSystem */
+/** @typedef {import("../../declarations/WebpackOptions").Externals} Externals */
+/** @typedef {import("../../declarations/plugins/DllReferencePlugin").DllReferencePluginOptions} DllReferencePluginOptions */
+/** @typedef {import("../../declarations/plugins/DllReferencePlugin").DllReferencePluginOptionsContent} DllReferencePluginOptionsContent */
+/** @typedef {import("../../declarations/plugins/DllReferencePlugin").DllReferencePluginOptionsManifest} DllReferencePluginOptionsManifest */
+/** @typedef {import("../Compiler")} Compiler */
+/** @typedef {import("../Compiler").CompilationParams} CompilationParams */
+/** @typedef {import("../util/fs").InputFileSystem} InputFileSystem */
 
 /** @typedef {{ path: string, data: DllReferencePluginOptionsManifest | undefined, error: Error | undefined }} CompilationDataItem */
 
@@ -41,14 +40,14 @@ class DllReferencePlugin {
 	apply(compiler) {
 		compiler.hooks.validate.tap(PLUGIN_NAME, () => {
 			compiler.validate(
-				() => require("../schemas/plugins/DllReferencePlugin.json"),
+				() => require("../../schemas/plugins/DllReferencePlugin.json"),
 				this.options,
 				{
 					name: "Dll Reference Plugin",
 					baseDataPath: "options"
 				},
 				(options) =>
-					require("../schemas/plugins/DllReferencePlugin.check")(options)
+					require("../../schemas/plugins/DllReferencePlugin.check")(options)
 			);
 		});
 		compiler.hooks.compilation.tap(
