+++ lib/dll/LibManifestPlugin.js
@@ -6,16 +6,16 @@
 "use strict";
 
 const asyncLib = require("neo-async");
-const EntryDependency = require("./dependencies/EntryDependency");
-const { someInIterable } = require("./util/IterableHelpers");
-const { compareModulesById } = require("./util/comparators");
-const { dirname, mkdirp } = require("./util/fs");
+const EntryDependency = require("../dependencies/EntryDependency");
+const { someInIterable } = require("../util/IterableHelpers");
+const { compareModulesById } = require("../util/comparators");
+const { dirname, mkdirp } = require("../util/fs");
 
-/** @typedef {import("./ChunkGraph").ModuleId} ModuleId */
-/** @typedef {import("./Compiler")} Compiler */
-/** @typedef {import("./Compiler").IntermediateFileSystem} IntermediateFileSystem */
-/** @typedef {import("./Module").BuildMeta} BuildMeta */
-/** @typedef {import("./ExportsInfo").ExportInfoName} ExportInfoName */
+/** @typedef {import("../ChunkGraph").ModuleId} ModuleId */
+/** @typedef {import("../Compiler")} Compiler */
+/** @typedef {import("../Compiler").IntermediateFileSystem} IntermediateFileSystem */
+/** @typedef {import("../Module").BuildMeta} BuildMeta */
+/** @typedef {import("../ExportsInfo").ExportInfoName} ExportInfoName */
 
 /**
  * Defines the manifest module data type used by this module.
