+++ lib/index.js
@@ -222,17 +222,8 @@ module.exports = mergeExports(fn, {
 	get DefinePlugin() {
 		return require("./DefinePlugin");
 	},
-	get DelegatedPlugin() {
-		return require("./DelegatedPlugin");
-	},
 	get Dependency() {
 		return require("./Dependency");
-	},
-	get DllPlugin() {
-		return require("./DllPlugin");
-	},
-	get DllReferencePlugin() {
-		return require("./DllReferencePlugin");
 	},
 	get DynamicEntryPlugin() {
 		return require("./DynamicEntryPlugin");
@@ -282,9 +273,6 @@ module.exports = mergeExports(fn, {
 			"webpack.JavascriptModulesPlugin has moved to webpack.javascript.JavascriptModulesPlugin",
 			"DEP_WEBPACK_JAVASCRIPT_MODULES_PLUGIN"
 		)();
-	},
-	get LibManifestPlugin() {
-		return require("./LibManifestPlugin");
 	},
 	get LibraryTemplatePlugin() {
 		return util.deprecate(
@@ -601,6 +589,35 @@ module.exports = mergeExports(fn, {
 		},
 		get EnableLibraryPlugin() {
 			return require("./library/EnableLibraryPlugin");
+		}
+	},
+
+	// TODO remove in webpack 6 in favor of `dll` scope
+	get DelegatedPlugin() {
+		return require("./dll/DelegatedPlugin");
+	},
+	get DllPlugin() {
+		return require("./dll/DllPlugin");
+	},
+	get DllReferencePlugin() {
+		return require("./dll/DllReferencePlugin");
+	},
+	get LibManifestPlugin() {
+		return require("./dll/LibManifestPlugin");
+	},
+
+	dll: {
+		get DelegatedPlugin() {
+			return require("./dll/DelegatedPlugin");
+		},
+		get DllPlugin() {
+			return require("./dll/DllPlugin");
+		},
+		get DllReferencePlugin() {
+			return require("./dll/DllReferencePlugin");
+		},
+		get LibManifestPlugin() {
+			return require("./dll/LibManifestPlugin");
 		}
 	},
