+++ lib/util/internalSerializables.js
@@ -181,7 +181,6 @@ module.exports = {
 
 	DelegatedModule: () => require("../DelegatedModule"),
 	DependenciesBlock: () => require("../DependenciesBlock"),
-	DllModule: () => require("../DllModule"),
 	ExternalModule: () => require("../ExternalModule"),
 	FileSystemInfo: () => require("../FileSystemInfo"),
 	InitFragment: () => require("../InitFragment"),
@@ -221,6 +220,9 @@ module.exports = {
 		require("../errors/CommentCompilationWarning"),
 	"errors/NodeStuffInWebError": () => require("../errors/NodeStuffInWebError"),
 	"errors/JSONParseError": () => require("../errors/JSONParseError"),
+
+	"dll/DelegatedModule": () => require("../dll/DelegatedModule"),
+	"dll/DllModule": () => require("../dll/DllModule"),
 
 	"util/LazySet": () => require("../util/LazySet"),
 	"util/registerExternalSerializer": () => {
