+++ types.d.ts
@@ -24456,6 +24456,14 @@ declare namespace exports {
 	export namespace library {
 		export { AbstractLibraryPlugin, EnableLibraryPlugin };
 	}
+	export namespace dll {
+		export {
+			DelegatedPlugin,
+			DllPlugin,
+			DllReferencePlugin,
+			LibManifestPlugin
+		};
+	}
 	export namespace container {
 		export const scope: <T>(
 			scope: string,
@@ -24747,10 +24755,7 @@ declare namespace exports {
 		ContextExclusionPlugin,
 		ContextReplacementPlugin,
 		DefinePlugin,
-		DelegatedPlugin,
 		Dependency,
-		DllPlugin,
-		DllReferencePlugin,
 		DynamicEntryPlugin,
 		DotenvPlugin,
 		EntryOptionPlugin,
@@ -24766,7 +24771,6 @@ declare namespace exports {
 		InitFragment,
 		IgnorePlugin,
 		JavascriptModulesPlugin,
-		LibManifestPlugin,
 		LibraryTemplatePlugin,
 		LoaderOptionsPlugin,
 		LoaderTargetPlugin,
@@ -24795,6 +24799,10 @@ declare namespace exports {
 		WebpackOptionsDefaulter,
 		ValidationError as WebpackOptionsValidationError,
 		ValidationError,
+		DelegatedPlugin,
+		DllPlugin,
+		DllReferencePlugin,
+		LibManifestPlugin,
 		EntryLibIndex as Entry,
 		EntryNormalized,
 		EntryObject,
